refactor(mcp-server): extract auth and observability helpers (no behavior change)
